# openapi/model-service-v2.yml
openapi: '3.0.0'
info:
  title: 'Model Service'
  description: 'Model Service v2 API - This API is under active development and subject to change.'
  version: '0.0.1'
servers:
  - url: https://api.pennsieve.io/models/v2
  - url: https://api.pennsieve.net/models/v2
paths:
  '/organizations/{organizationId}/datasets/{datasetId}/models':
    post:
      tags:
        - Models
      summary: 'Create a new model'
      description: 'Create a new model'
      operationId: server.api.v2.models.create_model
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
      requestBody:
        description: Model to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModel'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
      security:
        - Bearer: []
    get:
      tags:
        - Models
      summary: 'Get all the models for a dataset'
      description: 'Get all the models for a dataset'
      operationId: server.api.v2.models.get_all_models
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Model'
      security:
        - Bearer: []
  '/organizations/{organizationId}/datasets/{datasetId}/models/{modelIdOrName}':
    get:
      tags:
        - Models
      summary: 'Gets a model by ID'
      description: 'Gets a model by ID'
      operationId: server.api.v2.models.get_model
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID or name
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
      security:
        - Bearer: []

    put:
      tags:
        - Models
      summary: 'Update a model'
      description: 'Update a model'
      operationId: server.api.v2.models.update_model
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID or name
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModel'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
      security:
        - Bearer: []

    delete:
      tags:
        - Models
      summary: 'Delete a model'
      description: 'Delete a model'
      operationId: server.api.v2.models.delete_model
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID or name
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
      responses:
        "200":
          description: successful operation
      security:
        - Bearer: []
  '/organizations/{organizationId}/datasets/{datasetId}/models/{modelIdOrName}/properties':
    get:
      tags:
        - Models
      summary: 'Get all the properties for a model'
      description: 'Get all the properties for a model'
      operationId: server.api.v2.models.get_all_properties
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID or name
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
      security:
        - Bearer: []
    put:
      tags:
        - Models
      summary: 'Create or update or the properties for a model'
      description: 'Create or update or the properties for a model'
      operationId: server.api.v2.models.update_properties
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID or name
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateProperties'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
      security:
        - Bearer: []
  '/organizations/{organizationId}/datasets/{datasetId}/models/{modelId}/properties/{propertyName}':
    delete:
      tags:
        - Models
      summary: 'Delete a model property'
      description: 'Delete a model property'
      operationId: server.api.v2.models.delete_property
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelId
          in: path
          description: Model ID
          required: true
          schema:
            type: string
        - name: propertyName
          in: path
          description: Property name
          required: true
          schema:
            type: string
        - name: modifyRecords
          in: query
          description: Whether or not to modify existing records by removing this property
          required: false
          schema:
            type: boolean
            default: False
      responses:
        '200':
          description: successful operation
      security:
        - Bearer: []
...
  '/organizations/{organizationId}/datasets/{datasetId}/models/{modelIdOrName}/relationships':
    get:
      tags:
        - Models
        - Relationships
      summary: 'Get all outgoing relationships for given model'
      description: 'Get all outgoing relationships for given model'
      operationId: server.api.v2.models.get_relationships
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID or name
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelRelationship'
      security:
        - Bearer: []

    post:
      tags:
        - Models
        - Relationships
      summary: 'Create a new model relationship'
      description: 'Create a new model relationship'
      operationId: server.api.v2.models.create_relationship
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID or name
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
      requestBody:
        description: The model relationship to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelRelationship'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRelationship'
      security:
        - Bearer: []
  '/organizations/{organizationId}/datasets/{datasetId}/models/{modelIdOrName}/relationships/{relationshipId}':
    get:
      tags:
        - Models
        - Relationships
      summary: 'Get all outgoing relationships for given model'
      description: 'Get all outgoing relationships for given model'
      operationId: server.api.v2.models.get_relationship
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
        - name: relationshipId
          in: path
          description: Relationship ID
          required: true
          schema:
            $ref: '#/components/schemas/RelationshipIdentifier'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRelationship'
      security:
        - Bearer: []
    put:
      tags:
        - Models
        - Relationships
      summary: 'Update a model relationship'
      description: 'Update a model relationship'
      operationId: server.api.v2.models.update_relationship
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
        - name: relationshipId
          in: path
          description: Relationship ID
          required: true
          schema:
            $ref: '#/components/schemas/RelationshipIdentifier'
      requestBody:
        description: The model relationship to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateModelRelationship'
      responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ModelRelationship'
      security:
        - Bearer: []
    delete:
      tags:
        - Models
        - Relationships
      summary: 'Delete a model relationship'
      description: 'Delete a model relationship'
      operationId: server.api.v2.models.delete_relationship
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: modelIdOrName
          in: path
          description: Model ID
          required: true
          schema:
            $ref: '#/components/schemas/ModelIdentifier'
        - name: relationshipId
          in: path
          description: Relationship ID
          required: true
          schema:
            $ref: '#/components/schemas/RelationshipIdentifier'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: []

  # '/organizations/{organizationId}/datasets/{datasetId}/records/{recordId}/relationships/{relationshipType}':
  #   get:
  #     summary: Get all outgoing relations from a record.
  #     description: Get all outgoing relations from a record.
  #     operationId: server.api.v2.records.get_relationships
  #     parameters:
  #       - name: organizationId
  #         in: path
  #         description: Organization ID
  #         required: true
  #         schema:
  #           $ref: '#/components/schemas/OrganizationIdentifier'
  #             - name: datasetId
  #         in: path
  #         description: Dataset Id
  #         required: true
  #         schema:
  #           $ref: '#/components/schemas/DatasetIdentifier'
  #       - name: recordId
  #         in: path
  #         description: record Id
  #         required: true
  #         schema:
  #           type: string
  #       - name: relationshipType
  #         in: path
  #         description: relation type
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       '200':
  #         description: successful operation
  #         content:
  #           application/json:
  #             schema:
  #               type: array
  #               items:
  #                 $ref: '#/components/schemas/RecordRelationship'
  #     security:
  #       - Bearer: []
  '/organizations/{organizationId}/datasets/{datasetId}/records/{recordId}/relationships':
    get:
      tags:
        - Records
        - Relationships
      summary: 'Get all outgoing relations from a record'
      description: 'Get all outgoing relations from a record'
      operationId: server.api.v2.records.get_relationships
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: recordId
          in: path
          description: Record ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecordRelationship'
      security:
        - Bearer: []
    post:
      tags:
        - Records
        - Relationships
      summary: 'Create a relationship between records'
      description: 'Create a relationship between records'
      operationId: server.api.v2.records.create_relationship
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: recordId
          in: path
          description: Record ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: The record relationship to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordRelationship'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordRelationship'
      security:
        - Bearer: []
  '/organizations/{organizationId}/datasets/{datasetId}/records/{recordId}/relationships/{relationshipId}':
    get:
      tags:
        - Records
        - Relationships
      summary: 'Get an outgoing record relationship from a record'
      description: 'Get an outgoing record relationship from a record'
      operationId: server.api.v2.records.get_relationship
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: recordId
          in: path
          description: Record ID
          required: true
          schema:
            type: string
            format: uuid
        - name: relationshipId
          in: path
          description: Relation ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordRelationship'
      security:
        - Bearer: []
    delete:
      tags:
        - Records
        - Relationships
      summary: 'Delete an outgoing relationship from a record'
      description: 'Delete an outgoing relationship from a record'
      operationId: server.api.v2.records.delete_relationship
      parameters:
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: datasetId
          in: path
          description: Dataset ID
          required: true
          schema:
            $ref: '#/components/schemas/DatasetIdentifier'
        - name: recordId
          in: path
          description: Record ID
          required: true
          schema:
            type: string
            format: uuid
        - name: relationshipId
          in: path
          description: Relation ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: []
...
  '/organizations/{organizationId}/search/records':
    post:
      tags:
        - Search
      summary: 'Search records across datasets'
      operationId: server.api.v2.search.records
      parameters:
        - name: X-Bf-Trace-Id
          in: header
          schema:
            type: string
          required: true
        - name: organizationId
          in: path
          description: Organization ID
          required: true
          schema:
            $ref: '#/components/schemas/OrganizationIdentifier'
        - name: limit
          in: query
          description: limit
          required: false
          schema:
            type: integer
            format: int64
            default: 25
        - name: offset
          in: query
          description: offset
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: orderBy
          in: query
          description: Order by property
          required: false
          schema:
            type: string
            maxLength: 32
            nullable: true
        - name: orderDirection
          in: query
          description: Order by direction
          required: false
          schema:
            type: string
            enum: ["asc", "desc"]
            default: "asc"
      requestBody:
        description: The search request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordSearchResult'
      security:
        - Bearer: []
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: server.auth.decode_token
  schemas:
    OrganizationIdentifier:
      type: integer
...
    DatasetIdentifier:
      type: integer
...
    ModelIdentifier:
      type: string  # UUID or name
...
    RelationshipIdentifier:
      type: string
      format: uuid
...
    Model:
      type: object
      required:
        - id
        - createdAt
        - description
        - displayName
        - name
        - updatedAt
      properties:
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        id:
          type: string
          format: uuid
        count:
          type: integer
          format: int64
        propertyCount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateModel:
      type: object
      required:
        - name
        - displayName
        - description
      properties:
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
    Property:
      type: object
      required:
        - name
        - displayName
        - dataType
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        displayName:
          type: string
        dataType:
          $ref: '#/components/schemas/DataType'
        index:
          type: integer
          format: int64
        locked:
          type: boolean
        default:
          type: boolean
        modelTitle:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        description:
          type: string
        required:
          type: boolean
          default: false
    CreateOrUpdateProperties:
      type: array
      items:
        allOf:
          - $ref: '#/components/schemas/Property'
          - type: object
            required:
              # - id
              - name
              - displayName
              - dataType
...
    ModelRelationship:
      type: object
      required:
        - id
        - oneToMany
        - displayName
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        oneToMany:
          type: boolean
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateModelRelationship:
      type: object
      required:
        - type
        - to
        - oneToMany
      properties:
        "type":
          type: string
        to:
          type: string
          format: uuid
        oneToMany:
          type: boolean
          default: True
        displayName:
          type: string
          default: ""
    UpdateModelRelationship:
      type: object
      required:
        - displayName
      properties:
        displayName:
          type: string
    RecordRelationship:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
    CreateRecordRelationship:
      type: object
      required:
        - "type"
        - to
      properties:
        "type":
          type: string
        to:
          type: string
          format: uuid
...